refactor(memory): split memory.rs into submodules (#565)
Break the 1460-line `src/engine/memory.rs` into focused submodules so no
single file exceeds the 800-line guideline:

- `memory/cgroup.rs` — cgroup v1/v2 mount + memory limit parsing (napi-only)
- `memory/semaphore.rs` — fair `WeightedSemaphore` + `MemoryPermit`
- `memory/estimate.rs` — header parsing, BPP detection, peak-memory estimation
- `memory/mod.rs` — capacity sizing, global semaphore, `detect_available_memory`,
  and `calculate_memory_based_concurrency`

No public API changes: external callers reach the same symbols via
`memory::*` re-exports. Internal seams use `pub(super)`. `cgroup` is fully
`#[cfg(feature = "napi")]`.

The two cache tests are now serialized with a local mutex; previously they
were napi-gated and never ran in CI, so the race was invisible. Exposing
them under both feature sets is more valuable than re-gating them.
